requirements.txt
numpy
scipy
matplotlib
PySoundFile
pyAudio
psutil
keras
tensorflow
librosa
pillow
imageio